Move constants to subgraph-core
