+++ projects/subgraph-bean/src/BeanHandler.ts
@@ -1,7 +1,6 @@
-import { log } from "@graphprotocol/graph-ts";
 import { Transfer } from "../generated/Bean/Bean";
 import { loadBean, updateBeanSupplyPegPercent } from "./utils/Bean";
-import { ADDRESS_ZERO } from "./utils/Constants";
+import { ADDRESS_ZERO } from "../../subgraph-core/utils/Constants";
 
 export function handleTransfer(event: Transfer): void {
   if (event.params.from == ADDRESS_ZERO || event.params.to == ADDRESS_ZERO) {
